Set NB_USER to the user logging in
This did't actually help as much as I'd hoped. Ideally this would have
solved the problem of the notebook using /home/user for the home directory
instead of the old /home/jovyan. I can't seem to get that to work properly

Additionally, I set the storage to use a 'guest collection' instead of the
HA mapped collection from before.